Fix race condition during pubsub topic release.
